# Speed up passing ASCII-only strings to WASM