--- a/tests/headless/main.rs
+++ b/tests/headless/main.rs
@@ -50,3 +50,4 @@ pub fn import_export_same_name() {
 pub mod snippets;
 pub mod modules;
 pub mod anyref_heap_live_count;
+pub mod strings;