@alexcrichton 2019-04-18
Thanks for this! I was wondering though if we could perhaps be more strict about changes here to head off regressions like `encodeInto` slipped through previously? Would you be able to generate a benchmark that others could run as well? Is it possible to enhance the test suite in this regard and/or write node-facing benchmarks that we could test on CI? It'd be great to have more coverage of this since it does seem like the logic is getting especially tricky.

cc @fitzgen