@RReverser 2019-04-18
> Would you be able to generate a benchmark that others could run as well?

I think this would make sense, yeah, although for now I've been focusing on finishing serde-wasm-bindgen (which has own benchmarks presented above and relies on optimisations).

So yes, I think it's a great idea given the risks, but I'm not 100% sure when I'll get around to adding all these generic tests and benchmarks.

How about leaving this open for now and maybe someone else would want to add these string tests in the meanwhile?