@RReverser 2019-04-18
Although wait, in terms of tests - I think you said we have coverage now that newer Firefox runs on Azure?