@fitzgen 2019-04-19
Regarding benchmarks: I know you also had a micro benchmark for various into/from wasm ABI stuff, Alex. I think it would be useful to collect these in a single place where anyone can run them and we make sure that they build in our CI.

How about a `./benchmarks` directory that is essentially the same as `./examples` but for internal-facing [micro-]benchmarks rather than user-facing examples? (Note: I'm not suggesting `./benches` since that could potentially confuse cargo). Each `./benchmarks/foo` directory could have a little README describing what it is measuring and how to build+run the benchmark.