@alexcrichton 2019-04-22
Ah that's right, I did! After some digging I also apparently [even published the benchmark](https://alexcrichton.github.io/rust-wasm-benchmark/). 

I actually think it'd be pretty cool if we published the benchmark for our own internal usage from CI, and that way we could always redirect folks to take a look and compare numbers themselves. I can try to whip something up and make a PR