@RReverser 2019-04-30
> Sorry no I haven't had enough time to work through this and try to get something set up. In the meantime could you share how you were benchmarking this?

Alright, I've created isolated benchmark set for passing strings into WASM here, instructions included in README: https://github.com/RReverser/wasm-bindgen-string-benches

I've ran them myself in latest Firefox and Chrome to provide some numbers and, as expected, because these are isolated string benchmarks and not whole-app improvements like above, the difference is even bigger, especially for small ASCII strings, which are particularly common and where C++ overhead is particularly noticeable.


string / ops/s | Chrome (before) | Firefox (before) | Chrome (after) | Firefox (after) | Chrome (speed up) | Firefox (speed up)
-- | -- | -- | -- | -- | -- | --
`ja` | 616,791 | 1,740,501 | 6,940,567 | 10,975,125 | 11.3x | 6.3x
`aym0566x` | 641,746 | 1,715,014 | 4,207,093 | 9,756,924 | 6.6x | 5.7x
`505874924095815000` | 615,011 | 1,701,137 | 6,589,310 | 8,152,707 | 10.7x | 4.8x
`Sun Aug 31 00:29:15 +0000 2014` | 585,353 | 1,678,253 | 5,307,496 | 6,643,487 | 9.1x | 4.0x
`https://pbs.twimg.com/profile_images/497760886795153410/LDjAwR_y_normal.jpeg` | 523,348 | 1,492,255 | 2,940,477 | 3,937,804 | 5.6x | 2.6x
`<a href="http://twitter.com/download/iphone" rel="nofollow">Twitter for iPhone</a>` | 533,109 | 1,516,124 | 2,788,777 | 3,762,589 | 5.2x | 2.5x
`@aym0566x             名前:前田あゆみ...` | 299,053 | 622,844 | 258,450 | 966,838 | 0.9x | 1.6x

![image](https://user-images.githubusercontent.com/557590/56990757-cfa88780-6b8d-11e9-957c-18dc6f8b41d8.png)