@RReverser 2019-04-30
As for more tests - no, I didn't have time to work on that.