@RReverser 2019-05-10
> Thanks for the benchmark! I can indeed reproduce some similar performance gains on Firefox myself.

Hmm, I just found out that I left few mistakes in the benchmark code - did you manage to run the same benchmarks when you wrote that? Anyway, should be fixed now.