@alexcrichton 2019-05-14
Ok just wanted to make sure. It would also be helpful to write requested tests ahead of time for PRs!

This isn't a massive project where some breakage on master destroy's everyone's workflow. I'm simply extremely busy yesterday and today and it took a bit to merge a fix. No sweat if things are broken temporarily.